// src/types/tagTypes.ts
export enum tagTypes {
  user = "user",
  shop = "shop",
  otp = "otp",
  restaurant = "restaurant",
  menu = "menu",
  table = "table",
  booking = "booking",
  order = "order",
  notification = "notification",
  content = "content",
  wallet = "wallet",
  TopRestaurant = "TopRestaurant",
  income = "income",
  vendorIncome = "vendorIncome",
  vendorEarnings = "vendorEarnings",
  vendorOrders = "vendorOrders",
  products = "Products",
  product = "product",
  refund = "refund",
}

export const tagTypesList = [
  tagTypes.user,
  tagTypes.income,
  tagTypes.shop,
  tagTypes.otp,
  tagTypes.restaurant,
  tagTypes.menu,
  tagTypes.table,
  tagTypes.booking,
  tagTypes.order,
  tagTypes.notification,
  tagTypes.content,
  tagTypes.wallet,
  tagTypes.TopRestaurant,
  tagTypes.products,
  tagTypes.product,
  tagTypes.vendorIncome,
  tagTypes.vendorEarnings,
  tagTypes.vendorOrders,
  tagTypes.refund,
];